botbu/requirements.txt
requests
python-kucoin
